[Source-mysql] : Remove mysql initial source operations
