+++ airbyte-integrations/connectors/source-mysql/src/main/java/io/airbyte/integrations/source/mysql/MySqlSourceOperations.java
@@ -46,6 +46,7 @@ import com.mysql.cj.jdbc.result.ResultSetMetaData;
 import com.mysql.cj.result.Field;
 import io.airbyte.cdk.db.SourceOperations;
 import io.airbyte.cdk.db.jdbc.AbstractJdbcCompatibleSourceOperations;
+import io.airbyte.integrations.source.mysql.initialsync.CdcMetadataInjector;
 import io.airbyte.protocol.models.JsonSchemaType;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
@@ -54,6 +55,7 @@ import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.OffsetDateTime;
 import java.time.format.DateTimeParseException;
+import java.util.Optional;
 import java.util.Set;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -65,6 +67,28 @@ public class MySqlSourceOperations extends AbstractJdbcCompatibleSourceOperation
       SMALLINT_UNSIGNED, MEDIUMINT, MEDIUMINT_UNSIGNED, INT, INT_UNSIGNED, BIGINT, BIGINT_UNSIGNED,
       FLOAT, FLOAT_UNSIGNED, DOUBLE, DOUBLE_UNSIGNED, DECIMAL, DECIMAL_UNSIGNED, DATE, DATETIME, TIMESTAMP,
       TIME, YEAR, VARCHAR, TINYTEXT, TEXT, MEDIUMTEXT, LONGTEXT);
+
+  private final Optional<CdcMetadataInjector> metadataInjector;
+
+  public MySqlSourceOperations() {
+    super();
+    this.metadataInjector = Optional.empty();
+  }
+
+  public MySqlSourceOperations(final Optional<CdcMetadataInjector> metadataInjector) {
+    super();
+    this.metadataInjector = metadataInjector;
+  }
+
+  @Override
+  public JsonNode rowToJson(final ResultSet queryContext) throws SQLException {
+    final ObjectNode jsonNode = (ObjectNode) super.rowToJson(queryContext);
+    if (!metadataInjector.isPresent()) {
+      return jsonNode;
+    }
+    metadataInjector.get().inject(jsonNode);
+    return jsonNode;
+  }
 
   /**
    * @param colIndex 1-based column index.
